[posix] unify the RCP reset sequence
This commit unifies the RCP reset sequence on the HDLC and SPI interfaces.
The Thread stack host sends command `SPINEL_CMD_RESET` to reset the RCP
first. If the software reset fails, the host will use the hardware reset
to reset the RCP.

Reasons for unifing the RCP reset sequence:
- The parameters of the Thread satck are fixed on Android. The developer
  can't specify different paramerers to reset HDLC and SPI interfaces.

- If the Thread stack and the BLE stack are running on the same radio chip.
   Software reset the RCP first can minimize the impact on BLE stack.