[netdata-publisher] optimize route prefixes
This commit adds a new mechanism in `NetworkData::Publisher` to
optimize published route prefixes.

If three or more route prefixes (excluding NAT64 prefixes) share a
common sub-prefix of length 7, they are marked as optimized and
replaced with the compact sub-prefix. For example, three or more ULA
prefixes will be replaced with the `fc00::/7` route prefix. If the
number of prefixes drops below three, the compact prefix is removed
and the original prefixes are again published. The compact prefix is
added using the highest preference level among all prefixes that it
replaces.

`OPENTHREAD_CONFIG_NETDATA_PUBLISHER_OPTIMIZE_ROUTES` enables this
feature and is enabled by default.

This commit also adds `CompactFlag` in `HasRouteEntry` flags. This flag
is used to indicate that a route prefix is compact.

`test-020-network-data-publisher-compact-routes.py` is added testing
the behavior of the new feature. In particular, when many route
prefixes are published together or separately and we go over or below
the threshold of 3 prefixes. Also the case where the compact prefix
itself happens to be a published route is validated.