[net-diag] define Child, Neighbor, and Child IPv6 Address List TLVs
This commit contains multiple changes to Network Diagnostics
modules.

It adds three new TLVs: Child TLV, Child IPv6 Address List TLV, and
Neighbor TLV. Child TLV contains information about a child entry
including RLOC16, extended MAC address, mode (rx-on-idle, FTD/MTD,
full netdata, CSL-sync), timeout, age, supervision interval, CSL
period and timeout, and link quality info such as RSS (last and
average) and frame/message error rates. Neighbor TLV provides similar
info for a router neighbor. The Child IPv6 Address List TLV allows to
query a parent to get the list of IPv6 addresses registered by its
MTD children.

The new TLVs can be requested in Network Diagnostic Get Query "d/dq"
message to a router (unicast) or multiple routers (multicast). The
router responds with one or more Network Diagnostic Get Answer "d/da"
messages.

We allow multiple Answer messages to be sent in response to a Query.
This is to support the case where a router has many children, and we
want to avoid sending a large message that contains all of the TLVs
for all of the children. Instead, we can send a sequence of Answer
messages, each of which contains a subset of the TLVs.

Each Answer message includes an Answer TLV that indicates the index of
the message in the sequence. The first Answer message has an index of
0, the second has an index of 1, and so on. The Answer TLV also
indicates whether the message is the last one in the sequence, or if
there are more Answer messages to come. Answer messages are sent in
sequence, and the next one is not sent until we receive a successful
CoAP ACK/response for the previous one.

This commit adds a Query ID TLV (with a 16-bit identifier as it
value) that can be optionally included in a Query message. The Query
ID is echoed back in all Answer messages associated with the same
Query message. This allows a node that sends multiple Queries to
distinguish between the received Answers.

This commit updates `MeshDiag` to use the newly added TLVs
to query the child table or children IPv6 addresses of a router.
New public OpenThread APIs under `otMeshDiag` are added for this,
along with CLI commands and their documentation. A test-case is added
to validate the newly added mechanism.