[mac] Allow setting the timeout for receiving a Data Frame
This setting is targeted for SEDs which may need to decrease the timeout for
receiving a Data Frame (which follows an ACK with FP bit set) for power saving reasons.

IEEE 802.15.4/2015 specification doesn't enforce a specific timeout value for
this parameter so the application should be able to configure it.

Signed-off-by: Doru Gucea <doru-cristian.gucea@nxp.com>