Add posix vendor support with lib dependencies A vendor interface implementation for the posix library is currently supported through linking a specified c++ source file. However, this is unable to accomodate dependencies for external libraries or header files that this vendor implementation source file may need.
The posix vendor support is changed to allow a vendor to specify a cmake
RcpVendorDeps module file to resolve missing dependencies for a provided
vendor support source file.

Additionally, this commit adds the ability for a vendor to define a
radio url help message when the vendor rcp bus is selected instead of
defaulting to the UART help message.