[net-diag] define child IPv6 address list TLV
This commit adds Network Diagnostic Child IPv6 Address List TLV
(with TLV type 26). This TLV provides the list of IPv6 addresses
which an MTD child has registered with its parent.

This commit adds mechanism to a query to a parent to retrieve the
IPv6 addresses of all its MTD children using this TLV. `MeshDiag`
is updated to provide this functionality.