[posix] update upstream DNS addresses after a few minutes
On some platforms, the /etc/resolv.conf might be a generated file, and
is not available at the first few minutes after boot.

This PR records the time when we read the DNS server addresses from the
upstream, and reload the server list when it expires.

We may want to use inotify or add a platform specific API to notify the
openthread about the change of DNS server list.