[csl] update CSL public APIs to use milliseconds unit for period
This commit updates public APIs for getting and setting CSL period to
use milliseconds unit instead of the internal ten symbols unit. This
makes the APIs easier to use.

The CSL APIs have been renamed to follow the `otLinkGet/SetCsl{Item}()`
pattern, which is the common naming style of OpenThread. This
renaming will make the APIs more consistent and avoid potential
confusion with the now-removed APIs, which used a different unit for
CSL period.

This commit also updates the related CLI CSL commands:
- The `csl period` expects the given period to be in milliseconds.
- The `csl` command (which outputs all CSL parameter) shows the CSL
  period in millisecond unit (e.g., "Period: 160ms").

The NCP spinel `SPINEL_PROP_THREAD_CSL_PERIOD` is also updated to use
millisecond unit for CSL period.

The related test script are updated to use the new unit, in particular
the test harness `setCSLperiod()` in `harness-thci/OpenThread.py` is
updated (no need to convert from msec).

This commit also adds new helper methods to convert CSL period to and
from milliseconds and the ten symbols unit. When converting between
the two units, the value is rounded to the nearest integer. For
example, if the CSL period is 212 in ten symbols unit, which maps to
`212 * 0.160 = 33.92` milliseconds, the value of 34 milliseconds is
returned. Or if the target CSL period is 23 milliseconds, the
converted value will be 144 (in units of ten symbols) which maps to
`144 * 0.160 = 23.04` milliseconds instead of 143 which maps to 22.88
milliseconds.