[csl] new API to convert CSL period between msec and ten symbols unit
This commit adds new APIs to convert CSL period, which is specified in
units of ten symbols, to and from milliseconds.

When converting between units, the new APIs round to the nearest
value. For example, if the CSL period is `212` (in ten symbols unit)
which maps to `212 x 0.160 = 33.92` milliseconds, the value of `34`
milliseconds is returned. Or if the target CSL period is `23`
milliseconds, the converted value will be `144` (in units of ten
symbols) which maps to `144 x 0.160 = 23.04` milliseconds instead of
`143` which maps to `22.88` milliseconds.