[spinel] create spinel interface based on the radio url protocol
The posix platform is able to support the HDLC, SPI and vendor
spinel interfaces, but the spinel interface type can't be changed
dynamically. It is inconvenient to use different spinel interfaces
for Thread stack in Android. This commit enables the posix platform
to support the HDLC, SPI and vendor interface at the same time, and
the final spinel interface type is determined by the radio url protocol.

Some other changes:
<1> Not use CRTP style for the radio spinel.
<2> Deprecate the OT_POSIX_CONFIG_RCP_BUS and OPENTHREAD_POSIX_CONFIG_RCP_TIME_SYNC_INTERVAL.